+++ apps/web/app/search/page.tsx
@@ -1,4 +1,4 @@
-import SearchForm from "@/components/search-form";
+import SearchForm from "@/components/features/search-form";
 
 export default function SearchPage() {
   return (
