+++ apps/web/components/features/charts/cumulative-line-chart.tsx
