+++ apps/web/components/features/charts/item-count-bar-chart.tsx
