+++ apps/web/components/features/skeletons/stats-skeleton.tsx
