+++ apps/web/components/layouts/header-auth.tsx
@@ -1,6 +1,6 @@
 import { signOutAction } from "@/app/actions/auth";
 import Link from "next/link";
-import { Button } from "./ui/button";
+import { Button } from "@/components/ui/button";
 import { createClientForServer } from "@/utils/supabase/server";
 
 export default async function AuthButton() {
