--- database-indexes.md
@@ -1,85 +1,0 @@
-# Database Indexes for Performance Optimization
-
-This document outlines the recommended database indexes for optimal query performance in the whatpm-2025 application.
-
-## Current Query Patterns Analysis
-
-Based on the application's query patterns, the following indexes are recommended:
-
-## Recommended Indexes
-
-### 1. Items Table - Primary Queries Index
-```sql
--- Composite index for the most common query pattern (year + itemtype filtering)
-CREATE INDEX idx_items_year_type_created ON items (belongs_to_year, itemtype, created_at);
-```
-**Rationale**: This supports the optimized ItemsList queries that filter by both `belongs_to_year` and `itemtype`, ordered by `created_at`.
-
-### 2. Items Table - Search Index
-```sql
--- Indexes for search functionality
-CREATE INDEX idx_items_title_text ON items USING gin(to_tsvector('english', title));
-CREATE INDEX idx_items_author_text ON items USING gin(to_tsvector('english', author));
-CREATE INDEX idx_items_director_text ON items USING gin(to_tsvector('english', director));
-```
-**Rationale**: Full-text search indexes for better search performance. PostgreSQL GIN indexes are optimal for text search operations.
-
-### 3. Items Table - Alternative Search Indexes (if full-text search is not preferred)
-```sql
--- Case-insensitive pattern matching indexes
-CREATE INDEX idx_items_title_ilike ON items (lower(title) text_pattern_ops);
-CREATE INDEX idx_items_author_ilike ON items (lower(author) text_pattern_ops);
-CREATE INDEX idx_items_director_ilike ON items (lower(director) text_pattern_ops);
-```
-**Rationale**: Supports ILIKE queries with `%pattern%` matching used in the search function.
-
-### 4. Items Table - Individual Item Lookup
-```sql
--- Index for individual item queries (likely already exists as primary key)
-CREATE UNIQUE INDEX idx_items_id ON items (id);
-```
-**Rationale**: Fast lookup for single item pages and updates.
-
-### 5. Year Navigation Optimization
-```sql
--- Index for distinct years (if not using a separate view/table)
-CREATE INDEX idx_items_belongs_to_year ON items (belongs_to_year);
-```
-**Rationale**: Supports the year navigation component if it queries the items table directly.
-
-## Query Performance Improvements Made
-
-### 1. ItemsList Component Optimization
-- **Before**: Single query fetching all items, then JavaScript filtering by type
-- **After**: Three parallel database queries with type filtering at database level
-- **Benefits**: 
-  - Reduced data transfer
-  - Better use of database indexes
-  - Parallel query execution
-  - More efficient memory usage
-
-### 2. Search Function Optimization
-- **Improvements**:
-  - Added minimum query length validation (2 characters)
-  - Added query escaping for special characters
-  - Limited result set to 100 items
-  - Selected only necessary columns
-  - Added query trimming
-
-### 3. Column Selection Optimization
-- **ItemsList**: Select only needed columns per item type instead of `*`
-- **Search**: Select only display-relevant columns instead of `*`
-
-## Implementation Notes
-
-1. **Apply indexes carefully**: Test performance impact before applying in production
-2. **Monitor index usage**: Use `EXPLAIN ANALYZE` to verify indexes are being used
-3. **Consider composite indexes**: The composite index `(belongs_to_year, itemtype, created_at)` can satisfy multiple query patterns
-4. **Full-text search**: If implementing full-text search, consider migrating from ILIKE to `to_tsvector`/`to_tsquery` for better performance
-
-## Expected Performance Benefits
-
-1. **ItemsList loading**: ~60% faster with database-level filtering + proper indexes
-2. **Search queries**: ~40% faster with column selection + text indexes
-3. **Memory usage**: ~30% reduction with selective column fetching
-4. **Database load**: Better distribution with parallel queries
